% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactcharteditor.R
\name{ggplotly}
\alias{ggplotly}
\title{ggplotly}
\usage{
ggplotly(..., config = reactChartEditorConfiguration())
}
\arguments{
\item{...}{Aguments to pass on to plotly::ggplotly.}

\item{config}{The reactChartEditorConfiguration().}
}
\description{
A wrapper for ggplotly which specifies the default
react-chart-editor ui component to target, and includes the edit button in
the Plotly modebar.
}
