% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_editor.R
\name{messageReactChartEditor}
\alias{messageReactChartEditor}
\title{Message the React Chart Editor}
\usage{
messageReactChartEditor(session, inputId, value, configuration)
}
\arguments{
\item{session}{The session object; the current reactive domain.}

\item{inputId}{The HTML element id of the react-chart-editor to update.}

\item{value}{The value to pass to the editor through a message.}

\item{configuration}{The configuration to pass to the editor through a
message.}
}
\description{
Message the editor with a list of a value and a configuration.
}
