% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_editor.R
\name{plotly_editor}
\alias{plotly_editor}
\title{Plotly chart editor input control}
\usage{
plotly_editor(inputId)
}
\arguments{
\item{inputId}{The Shiny ID of the UI component.}
}
\description{
Create a Plotly chart editor input control.
}
