% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactcharteditor.R
\name{reactChartEditor}
\alias{reactChartEditor}
\title{React Chart Editor}
\usage{
reactChartEditor(id = "react-chart-editor")
}
\arguments{
\item{id}{The HTML element id of the react-chart-editor UI component being
created.}
}
\description{
Create a Plotly react-chart-editor component UI element. Running
this function makes your application take a dependency on the
react-chart-editor React component, retrieved through the jsdelivr CDN as a
side effect. It uses reactR to create a Shiny Input from the component,
returning the HTML that can be included anywhere in your application's UI
code.
}
