% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactcharteditor.R
\name{reactChartEditorConfiguration}
\alias{reactChartEditorConfiguration}
\title{React Chart Editor Configuration}
\usage{
reactChartEditorConfiguration(...)
}
\arguments{
\item{...}{Arguments to pass along to plotly::config.}
}
\description{
Set or retrieve the configuration object for your Plotly charts.
This wraps plotly::config to include the editor button, and provides the
opportunity to set a different event name for the edit button. If it is
never called explicitly the default name will be used and the only
customization to the Plotly configuration will be the inclusion of the edit
button.
}
