% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_editor.R
\name{ggplotly}
\alias{ggplotly}
\title{ggplotly}
\usage{
ggplotly(...)
}
\arguments{
\item{...}{Aguments to pass on to plotly::ggplotly}
}
\description{
A wrapper for ggplotly which specifies the default
react-chart-editor ui component to target, and includes the edit button in
the Plotly modebar.
}
