% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_editor.R
\name{reactChartEditor}
\alias{reactChartEditor}
\title{React Chart Editor}
\usage{
reactChartEditor(id, editButtonEventName)
}
\arguments{
\item{id}{The HTML element id of the react-chart-editor to create.}

\item{editButtonEventName}{The name of the input event that will be available
under input$editButtonEventName, for example.}
}
\description{
Create a Plotly react-chart-editor component UI element. Running
this function makes your application take a dependency on the
react-chart-editor React component, retrieved through the jsdelivr CDN as a
side effect. It uses reactR to create a Shiny Input from the component,
returning the HTML that can be included anywhere in your application's UI
code.
}
